pyproject.toml
[tool.poetry]
name = "telebotmemo"
version = "0.1.0"
description = "telegram bot for memoization"
authors = ["egor nikitin <eanikitin90gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
redis = "^4.6.0"
psycopg2 = "^2.9.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
